requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
magic-filter==1.0.9
multidict==6.0.4
nltk==3.8.1
openai==0.27.2
packaging==23.0
pluggy==1.0.0
pytest==7.2.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
tiktoken==0.3.2
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.14
yarl==1.8.2
